requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
torch==2.0.0
torchvision==0.15.1
Pillow==9.5.0
python-multipart==0.0.6
pytest==7.3.1
httpx==0.24.0
